+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v0.6.0
 Library:
 1. [feat] 新增类 kotonebot.interop.win.ShakeMouse，支持检测晃动鼠标检测动作，可以用于实现晃动鼠标自动停止脚本执行的功能
+2. [feat] 移除 kotonebot.backend.core.Image 类，新编写代码应该使用 kotonebot.primitives.Image。新的 Image 类暂时保留与原类一直的字段以向后兼容
 
 ## v0.5.0
 Library：